refactor(workspace)!: collapse markdown helpers under the materializer
Move the four document/markdown/* files into the materializer's own
folder so a new reader sees one owner: the materializer is the only
consumer, and the public package subpath was already
@epicenter/workspace/document/materializer/markdown. Rename markdown.ts
to assemble-markdown.ts to avoid the folder-name collision.

Drop the defineMutation wrap from push/pull/rebuild. Zero callers
outside the package routed through the action system, so the wrap added
ceremony (empty Type.Object schema, title/description, the trailing
{} argument every test had to type) without buying anything. The
builder now exposes plain async methods:

  before                              after
  markdown.push({})                   markdown.push()
  markdown.pull({})                   markdown.pull()
  markdown.rebuild({})                markdown.rebuild()
  markdown.rebuild({ table: 'x' })    markdown.rebuild('x')

Apps wanting these as CLI-visible actions can re-wrap them in their own
defineActions registry in one line per method.

BREAKING CHANGE: push/pull/rebuild on the markdown materializer builder
are no longer defineMutation actions. Callers that invoked them as
actions (e.g. markdown.push({})) must drop the empty-object argument
and switch markdown.rebuild({ table }) to markdown.rebuild(table). No
in-tree app needed any update; only tests and a future re-wrap path.
