+++ packages/workspace/src/document/materializer/markdown/index.ts
@@ -1,11 +1,7 @@
-export { assembleMarkdown } from '../../markdown/markdown.js';
-export { parseMarkdownFile } from '../../markdown/parse-markdown-file.js';
-export { prepareMarkdownFiles } from '../../markdown/prepare-markdown-files.js';
-export {
-	slugFilename,
-	toIdFilename,
-	toSlugFilename,
-} from '../../markdown/serializers.js';
+export { assembleMarkdown } from './assemble-markdown.js';
+export { parseMarkdownFile } from './parse-markdown-file.js';
+export { prepareMarkdownFiles } from './prepare-markdown-files.js';
+export { slugFilename, toIdFilename, toSlugFilename } from './serializers.js';
 export {
 	attachMarkdownMaterializer,
 	MaterializerPushError,
