+++ packages/workspace/src/document/materializer/markdown/materializer.test.ts
@@ -26,7 +26,7 @@ import {
 	createDisposableCache,
 	defineTable,
 } from '../../../index.js';
-import { parseMarkdownFile } from '../../markdown/parse-markdown-file.js';
+import { parseMarkdownFile } from './parse-markdown-file.js';
 import {
 	attachMarkdownMaterializer,
 	type MarkdownShape,
@@ -141,7 +141,7 @@ describe('push', () => {
 			'---\nid: post-2\ntitle: Draft Post\npublished: false\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(2);
 		expect(result.skipped).toBe(0);
@@ -166,7 +166,7 @@ describe('push', () => {
 		await writeTestFile('posts/readme.txt', 'not a markdown file');
 		await writeTestFile('posts/data.json', '{"id": "test"}');
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(1);
 		expect(result.skipped).toBe(0);
@@ -185,7 +185,7 @@ describe('push', () => {
 			'# Just a heading\n\nSome content\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(1);
 		expect(result.skipped).toBe(1);
@@ -196,7 +196,7 @@ describe('push', () => {
 	test('silently skips tables whose directories do not exist', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
 		// Don't create the posts directory. It should not exist.
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(0);
 		expect(result.skipped).toBe(0);
@@ -215,7 +215,7 @@ describe('push', () => {
 			'---\nid: post-bad\ntitle: 42\npublished: true\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(0);
 		expect(result.skipped).toBe(0);
@@ -250,7 +250,7 @@ describe('push', () => {
 			'---\nid: note-boom\n---\n\nirrelevant\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.errored).toBe(1);
 		const errorEvent = result.events.find((e) => e.kind === 'error');
@@ -275,7 +275,7 @@ describe('push', () => {
 			'---\nid: p3\ntitle: 42\npublished: true\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(1);
 		expect(result.skipped).toBe(1);
@@ -308,7 +308,7 @@ describe('push', () => {
 			'---\nid: note-1\n---\n\nThis is the body content\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(1);
 
@@ -327,7 +327,7 @@ describe('push', () => {
 			'---\nid: p1\ntitle: Hello\npublished: false\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(1);
 		expect(workspace.tables.posts.has('p1')).toBe(true);
@@ -343,7 +343,7 @@ describe('push', () => {
 			'---\nid: p1\ntitle: Original\npublished: false\n_v: 1\n---\n',
 		);
 
-		const first = await workspace.materializer.push({});
+		const first = await workspace.materializer.push();
 		expect(first.imported).toBe(1);
 
 		const { data: originalPost } = workspace.tables.posts.get('p1');
@@ -358,7 +358,7 @@ describe('push', () => {
 			'---\nid: p1\ntitle: Updated From Disk\npublished: true\n_v: 1\n---\n',
 		);
 
-		const second = await workspace.materializer.push({});
+		const second = await workspace.materializer.push();
 		expect(second.imported).toBe(1);
 
 		const { data: updatedPost } = workspace.tables.posts.get('p1');
@@ -379,7 +379,7 @@ describe('push', () => {
 			'---\nid: n1\nbody: Note body\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.push({});
+		const result = await workspace.materializer.push();
 
 		expect(result.imported).toBe(2);
 		expect(workspace.tables.posts.has('p1')).toBe(true);
@@ -409,7 +409,7 @@ describe('pull', () => {
 			_v: 1,
 		});
 
-		const result = await workspace.materializer.pull({});
+		const result = await workspace.materializer.pull();
 
 		expect(result.written).toBe(2);
 
@@ -432,7 +432,7 @@ describe('pull', () => {
 			_v: 1,
 		});
 
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 
 		const entries = await listTestDir('posts');
 		expect(entries).toContain('p1.md');
@@ -457,7 +457,7 @@ describe('pull', () => {
 		});
 		workspace.tables.notes.set({ id: 'n1', body: 'Custom body', _v: 1 });
 
-		const result = await workspace.materializer.pull({});
+		const result = await workspace.materializer.pull();
 
 		expect(result.written).toBe(1);
 
@@ -478,7 +478,7 @@ describe('pull', () => {
 			_v: 1,
 		});
 
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 
 		const entries = await listTestDir('blog');
 		expect(entries).toContain('p1.md');
@@ -488,7 +488,7 @@ describe('pull', () => {
 
 	test('writes nothing when table is empty', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
-		const result = await workspace.materializer.pull({});
+		const result = await workspace.materializer.pull();
 
 		expect(result.written).toBe(0);
 
@@ -505,7 +505,7 @@ describe('pull', () => {
 		});
 		workspace.tables.notes.set({ id: 'n1', body: 'Note', _v: 1 });
 
-		const result = await workspace.materializer.pull({});
+		const result = await workspace.materializer.pull();
 
 		expect(result.written).toBe(2);
 
@@ -533,7 +533,7 @@ describe('rebuild', () => {
 			published: true,
 			_v: 1,
 		});
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 		await writeTestFile(
 			'posts/orphan.md',
 			'---\nid: orphan\ntitle: Orphan\npublished: false\n_v: 1\n---\n',
@@ -543,7 +543,7 @@ describe('rebuild', () => {
 		expect(before).toContain('p1.md');
 		expect(before).toContain('orphan.md');
 
-		const result = await workspace.materializer.rebuild({});
+		const result = await workspace.materializer.rebuild();
 
 		expect(result.deleted).toBe(2); // p1.md + orphan.md both unlinked
 		expect(result.written).toBe(1); // only p1 re-written
@@ -564,13 +564,13 @@ describe('rebuild', () => {
 			_v: 1,
 		});
 		workspace.tables.notes.set({ id: 'n1', body: 'Note', _v: 1 });
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 		await writeTestFile(
 			'notes/orphan.md',
 			'---\nid: x\nbody: gone\n_v: 1\n---\n',
 		);
 
-		const result = await workspace.materializer.rebuild({ table: 'posts' });
+		const result = await workspace.materializer.rebuild('posts');
 
 		expect(result.deleted).toBe(1); // p1.md
 		expect(result.written).toBe(1); // p1 re-written
@@ -585,7 +585,7 @@ describe('rebuild', () => {
 	test('throws on unknown table name', async () => {
 		const { workspace } = await setup({ tables: (t) => [{ table: t.posts }] });
 		await expect(
-			workspace.materializer.rebuild({ table: 'notAThing' }),
+			workspace.materializer.rebuild('notAThing'),
 		).rejects.toThrow(/not in the materialized table set/);
 
 		workspace[Symbol.dispose]();
@@ -606,13 +606,13 @@ describe('rebuild', () => {
 			_v: 1,
 		});
 
-		const first = await workspace.materializer.rebuild({});
+		const first = await workspace.materializer.rebuild();
 		const stateAfterFirst = await listTestDir('posts');
 		const contentsAfterFirst = await Promise.all(
 			stateAfterFirst.map((f) => readTestFile(`posts/${f}`)),
 		);
 
-		const second = await workspace.materializer.rebuild({});
+		const second = await workspace.materializer.rebuild();
 		const stateAfterSecond = await listTestDir('posts');
 		const contentsAfterSecond = await Promise.all(
 			stateAfterSecond.map((f) => readTestFile(`posts/${f}`)),
@@ -671,7 +671,7 @@ describe('round-trip', () => {
 			_v: 1,
 		});
 
-		await workspace1.materializer.pull({});
+		await workspace1.materializer.pull();
 		workspace1[Symbol.dispose]();
 
 		// Verify files on disk have valid frontmatter
@@ -701,7 +701,7 @@ describe('round-trip', () => {
 		const workspace2 = cache2.open('test.roundtrip.2');
 		await workspace2.whenReady;
 
-		const result = await workspace2.materializer.push({});
+		const result = await workspace2.materializer.push();
 		expect(result.imported).toBe(2);
 
 		const { data: p1 } = workspace2.tables.posts.get('p1');
@@ -746,7 +746,7 @@ describe('round-trip', () => {
 		workspace.tables.notes.set(original);
 
 		// Pull to disk, then push from disk into a fresh workspace.
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 
 		// Also assert the pure inverse identity: fromMarkdown(toMarkdown(x)) ≡ x
 		expect(fromMarkdownFn(toMarkdownFn(original))).toEqual(original);
@@ -788,7 +788,7 @@ describe('round-trip', () => {
 		const original = { id: 'n1', body: 'Body content here', _v: 1 as const };
 		workspace.tables.notes.set(original);
 
-		await workspace.materializer.pull({});
+		await workspace.materializer.pull();
 
 		const disk = await readTestFile('notes/n1.md');
 		const parsed = parseMarkdownFile(disk);
